# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Toy Mall Inventory Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bills/checkout:
    post:
      operationId: checkout
      tags: [bills]
      summary: Create a new bill at checkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "201":
          description: Bill created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CheckoutItem:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
        price:
          type: number
        mrp:
          type: number
      required:
        - productId
        - quantity
        - price
    CheckoutInput:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CheckoutItem"
        paymentMode:
          type: string
          enum: [cash, upi, credit]
          description: >-
            Mode of payment. `credit` records the bill as unpaid (receivable);
            requires `customerPhone` so the debtor is identifiable.
        customerName:
          type: string
        customerPhone:
          type: string
        discount:
          type: number
          description: Raw discount value (e.g. 10 for 10% or 200 for ₹200)
        discountType:
          type: string
          enum: [percent, amount]
          description: How to interpret the discount value
      required:
        - items
        - paymentMode
    Bill:
      type: object
      properties:
        id:
          type: string
        billNumber:
          type: integer
        totalAmount:
          type: number
        amountPaid:
          type: number
          description: Cumulative amount collected against this bill.
        paymentStatus:
          type: string
          enum: [paid, partial, unpaid]
        itemsCount:
          type: integer
        paymentMode:
          type: string
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        discount:
          type: ["number", "null"]
        discountType:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - totalAmount
        - amountPaid
        - paymentStatus
        - itemsCount
        - paymentMode
        - createdAt
    BillItem:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        productName:
          type: string
        productSku:
          type: string
        quantity:
          type: integer
        price:
          type: number
        mrp:
          type: ["number", "null"]
        subtotal:
          type: number
      required:
        - id
        - productId
        - productName
        - productSku
        - quantity
        - price
        - subtotal
...
    CheckoutResponse:
      type: object
      properties:
        bill:
          $ref: "#/components/schemas/Bill"
        items:
          type: array
          items:
            $ref: "#/components/schemas/BillItem"
      required:
        - bill
        - items
...